+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the "amiga-debug" extension will be documented in this fi
 
 ## 1.5.6
 - FIX: disassembly: fix divul, divsl, FPU, fmoveml (see `68k-dis.ts`)
+- FIX: [#128 Can't use the image converter link ](https://github.com/BartmanAbyss/vscode-amiga-debug/issues/128)
 
 ## 1.5.5
 - NEW: VASM assembler with debugging support ([PR#123 by merry^Architect](https://github.com/BartmanAbyss/vscode-amiga-debug/pull/123))